search

+++ lib/ui/pages/add_product_page/add_product_models/add_product_models_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart' hide State;
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:get_it/get_it.dart';
-import 'package:samo_techno_crm/models/product_model/product_model.dart';
+import 'package:samo_techno_crm/models/cart_product/cart_product_model.dart';
 import 'package:samo_techno_crm/ui/pages/add_product_page/add_product_bloc.dart';
 import 'package:samo_techno_crm/ui/pages/add_product_page/add_product_event.dart';
 import 'package:samo_techno_crm/ui/pages/add_product_page/add_product_state.dart';
@@ -239,7 +239,7 @@ class AddProductModelsPage extends StatelessWidget {
                                 .children?[index].name,
                             categoryName: args["category_item_name"],
                             price: bloc.costCtrl.text.isNotEmpty
-                                ? int.parse(
+                                ? double.parse(
                                     bloc.costCtrl.text,
                                   )
                                 : 0,
